Implement ComputeRateCollection & StorageRateCollection

+++ cfme/tests/containers/test_chargeback.py
@@ -10,7 +10,6 @@ from humanfriendly import tokenize
 
 from cfme.containers.provider import ContainersProvider
 from cfme.intelligence.chargeback import assignments
-from cfme.intelligence.chargeback import rates
 from cfme.utils.log import logger
 from cfme.utils.units import CHARGEBACK_HEADER_NAMES
 from cfme.utils.units import parse_number
@@ -176,10 +175,11 @@ def compute_rate(appliance, rate_type, interval):
                              'fixed_rate': 1,
                              'variable_rate': variable_rate}
     }
-    ccb = rates.ComputeRate(description, fields=data, appliance=appliance)
-    ccb.create()
+    ccb = appliance.collections.compute_rates.create(description, fields=data)
     yield ccb
-    ccb.delete()
+
+    if ccb.exists:
+        ccb.delete()
 
 
 @pytest.fixture(scope='module')
